NON-WORKING COMMIT Moved cross sections and tools to separate modules
